+++ dashboard/.dockerignore
@@ -2,8 +2,8 @@
 *.png
 *.sh
 settings_template.json
-dashboard.service
-readme.md
+*.service
+*.md
 Dockerfile
 node_modules
 npm-debug.log
