+++ dashboard/Dockerfile
@@ -1,10 +1,7 @@
 # Puppetter image
-FROM node
+FROM node:11.10.1-stretch
 
-# Application directory
-WORKDIR /usr/src/app
-
-# Node install packages
+# Installing pdfjam and chromium dependecies
 # https://github.com/GoogleChrome/puppeteer/blob/master/.ci/node8/Dockerfile.linux
 RUN apt-get update && \
     apt-get -y install pdfjam xvfb gconf-service libasound2 libatk1.0-0 libc6 libcairo2 libcups2 \
@@ -13,14 +10,20 @@ RUN apt-get update && \
       libxcomposite1 libxcursor1 libxdamage1 libxext6 libxfixes3 libxi6 libxrandr2 libxrender1 libxss1 \
       libxtst6 ca-certificates fonts-liberation libappindicator1 libnss3 lsb-release xdg-utils wget && \
     rm -rf /var/lib/apt/lists/*
-COPY package.json ./
-RUN npm install
-
-# Copy all files
-COPY . .
 
 # Setting timezone
 ENV TZ=America/Montreal
 RUN ln -snf /usr/share/zoneinfo/$TZ /etc/localtime && echo $TZ > /etc/timezone
 
+# Application directory
+WORKDIR /usr/src/app
+
+# Node install packages
+COPY package.json ./
+RUN npm install
+
+# Copy all apprlication files add unwanted files .dockerignore
+COPY . .
+
+# Default application
 CMD ["node", "scheduledRender.js"]
