--- dashboard/build.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-groups
-docker build -t securbot/dashboard .
