+++ dashboard/package.json
@@ -2,7 +2,7 @@
   "name": "dashboard",
   "version": "1.0.0",
   "description": "export jira dashboard using chrome",
-  "main": "dashboard.js",
+  "main": "src/dashboard.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
