--- dashboard/render.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-docker run -v /var/www/html/dashboard:/usr/src/app/out securbot/dashboard node render
