+++ dashboard/scheduledRender.js
@@ -1,6 +1,6 @@
-const dModule = require('./dashboard.js');
+const dModule = require('./src/dashboard.js');
 const schedule = require('node-schedule');
-const mail = require('./mail.js');
+const mail = require('./src/mail.js');
 
 // Loading settings
 try{
