+++ dashboard/src/BigPic/manifest.json
