+++ dashboard/src/BigPic/node-big-pic.js
