--- dashboard/start.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-docker run -v /var/www/html/dashboard:/usr/src/app/out --restart unless-stopped -d securbot/dashboard
